testing: move config, chroot, rootfs out of testing.vmcore
For a while testing.vm.config has contained configuration elements that
are not specific to the VM testing architecture, but are instead useful
for all tests.

Similarly, the rootfs/chroot modules can now be used by the vmcore
tests, so they should not be bundled within testing.vm either.
